rename PeerRecord files to PeerManager
