+++ src/overlay/PeerManager.cpp
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "overlay/PeerRecord.h"
+#include "overlay/PeerManager.h"
 #include "lib/util/format.h"
 #include "main/Application.h"
 #include "overlay/StellarXDR.h"
