+++ src/overlay/PeerManagerTests.cpp
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "PeerRecord.h"
+#include "PeerManager.h"
 #include "database/Database.h"
 #include "lib/catch.hpp"
 #include "main/Application.h"
